# docs/openapi/openapi.json
{"openapi":"3.1.0","info":{"title":"MindOra API","description":"Public, admin, and RAG APIs for MindOra.","version":"v1"},"servers":[{"url":"http://localhost","description":"Generated server url"}],"paths":{"/api/v1/admin/tags/{id}":{"put":{"tags":["blog-admin-controller"],"operationId":"updateTag","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NameRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseTagView"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["blog-admin-controller"],"operationId":"deleteTag","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/admin/categories/{id}":{"put":{"tags":["blog-admin-controller"],"operationId":"updateCategory","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NameRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseCategoryView"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["blog-admin-controller"],"operationId":"deleteCategory","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/admin/articles/{id}":{"put":{"tags":["blog-admin-controller"],"operationId":"updateArticle","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseArticleView"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","parameters":[{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthView"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","parameters":[{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthView"}}}}}}},"/api/v1/admin/tags":{"get":{"tags":["blog-admin-controller"],"operationId":"listTags","parameters":[{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListTagView"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["blog-admin-controller"],"operationId":"createTag","parameters":[{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NameRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseTagView"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/admin/categories":{"get":{"tags":["blog-admin-controller"],"operationId":"listCategories","parameters":[{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListCategoryView"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["blog-admin-controller"],"operationId":"createCategory","parameters":[{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NameRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseCategoryView"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/admin/assets":{"get":{"tags":["asset-controller"],"operationId":"list","parameters":[{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListAssetView"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["asset-controller"],"operationId":"upload","parameters":[{"name":"assetType","in":"query","required":false,"schema":{"type":"string"}},{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAssetView"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/admin/articles":{"get":{"tags":["blog-admin-controller"],"operationId":"listArticles","parameters":[{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListArticleView"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["blog-admin-controller"],"operationId":"createArticle","parameters":[{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseArticleView"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/admin/articles/{id}/unpublish":{"post":{"tags":["blog-admin-controller"],"operationId":"unpublish","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseArticleView"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/admin/articles/{id}/publish":{"post":{"tags":["blog-admin-controller"],"operationId":"publish","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseArticleView"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/public/tags":{"get":{"tags":["blog-public-taxonomy-controller"],"operationId":"tags","parameters":[{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListTagView"}}}}}}},"/api/v1/public/health":{"get":{"tags":["health-controller"],"operationId":"health","parameters":[{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMapStringString"}}}}}}},"/api/v1/public/categories":{"get":{"tags":["blog-public-taxonomy-controller"],"operationId":"categories","parameters":[{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListCategoryView"}}}}}}},"/api/v1/public/assets/{id}":{"get":{"tags":["asset-controller"],"operationId":"content","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"byte"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/public/articles":{"get":{"tags":["blog-public-controller"],"operationId":"list_1","parameters":[{"name":"categoryId","in":"query","required":false,"schema":{"type":"string"}},{"name":"tagId","in":"query","required":false,"schema":{"type":"string"}},{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListArticleView"}}}}}}},"/api/v1/public/articles/{slug}":{"get":{"tags":["blog-public-controller"],"operationId":"detail","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}},{"name":"X-Trace-Id","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseArticleView"}}}}}}}},"components":{"schemas":{"NameRequest":{"type":"object","properties":{"name":{"type":"string","minLength":1}},"required":["name"]},"ApiResponseTagView":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/TagView"},"traceId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"TagView":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"ApiResponseCategoryView":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/CategoryView"},"traceId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"CategoryView":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"ArticleRequest":{"type":"object","properties":{"title":{"type":"string"},"slug":{"type":"string"},"summary":{"type":"string"},"body":{"type":"string"},"coverAssetId":{"type":"string"},"categoryId":{"type":"string"},"tagIds":{"type":"array","items":{"type":"string"},"uniqueItems":true},"visibility":{"type":"string"}}},"ApiResponseArticleView":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/ArticleView"},"traceId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"ArticleView":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"slug":{"type":"string"},"summary":{"type":"string"},"body":{"type":"string"},"coverAssetId":{"type":"string"},"categoryId":{"type":"string"},"tagIds":{"type":"array","items":{"type":"string"},"uniqueItems":true},"status":{"type":"string"},"visibility":{"type":"string"},"readCount":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"publishedAt":{"type":"string","format":"date-time"}}},"AuthRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"ApiResponseAuthView":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/AuthView"},"traceId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"AuthView":{"type":"object","properties":{"userId":{"type":"string"},"email":{"type":"string"},"accessToken":{"type":"string"}}},"ApiResponseAssetView":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/AssetView"},"traceId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"AssetView":{"type":"object","properties":{"id":{"type":"string"},"fileName":{"type":"string"},"mimeType":{"type":"string"},"size":{"type":"integer","format":"int64"},"assetType":{"type":"string"},"publicUrl":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"ApiResponseListTagView":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/TagView"}},"traceId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"ApiResponseMapStringString":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"object","additionalProperties":{"type":"string"}},"traceId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"ApiResponseListCategoryView":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/CategoryView"}},"traceId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"ApiResponseListArticleView":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/ArticleView"}},"traceId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"ApiResponseListAssetView":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/AssetView"}},"traceId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"ApiResponseVoid":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"data":{},"traceId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}